+++ packages/react-native-worklets/Common/cpp/worklets/WorkletRuntime/WorkletRuntimeDecorator.cpp
@@ -62,6 +62,7 @@ void WorkletRuntimeDecorator::decorate(
     const std::string &name,
     const std::shared_ptr<JSScheduler> &jsScheduler,
     const bool isDevBundle,
+    const bool bundleModeEnabled,
     const bool enableMicrotaskQueue,
     jsi::Object &&jsiWorkletsModuleProxy,
     const std::shared_ptr<EventLoop> &eventLoop,
@@ -75,7 +76,9 @@ void WorkletRuntimeDecorator::decorate(
 
   rt.global().setProperty(rt, RuntimeData::runtimeNameBindingName, jsi::String::createFromAscii(rt, name));
 
-  rt.global().setProperty(rt, "__DEV__", isDevBundle);
+  if (!bundleModeEnabled) {
+    rt.global().setProperty(rt, "__DEV__", isDevBundle);
+  }
 
   rt.global().setProperty(rt, "__workletsModuleProxy", std::move(jsiWorkletsModuleProxy));